+++ cocofest/models/ding2007/ding2007_with_fatigue.py
@@ -1,6 +1,6 @@
 from typing import Callable
 
-from casadi import MX, vertcat, tanh
+from casadi import MX, vertcat
 import numpy as np
 
 from bioptim import (
@@ -9,109 +9,113 @@ from bioptim import (
     NonLinearProgram,
     OptimalControlProgram,
 )
-from .ding2003 import DingModelFrequency
-from .state_configure import StateConfigure
+from cocofest.models.ding2007 import DingModelPulseWidthFrequency
+from cocofest.models.state_configure import StateConfigure
 
 
-class DingModelPulseIntensityFrequency(DingModelFrequency):
+class DingModelPulseWidthFrequencyWithFatigue(DingModelPulseWidthFrequency):
     """
     This is a custom models that inherits from bioptim. CustomModel.
     As CustomModel is an abstract class, some methods are mandatory and must be implemented.
     Such as serialize, name_dof, nb_state.
 
-    This is the Hmed 2018 model using the stimulation frequency and pulse intensity in input.
+    This is the Ding 2007 model using the stimulation frequency and pulse width in input.
 
-    Hmed, A. B., Bakir, T., Garnier, Y. M., Sakly, A., Lepers, R., & Binczak, S. (2018).
-    An approach to a muscle force model with force-pulse amplitude relationship of human quadriceps muscles.
-    Computers in Biology and Medicine, 101, 218-228.
+    Ding, J., Chou, L. W., Kesar, T. M., Lee, S. C., Johnston, T. E., Wexler, A. S., & Binder‐Macleod, S. A. (2007).
+    Mathematical model that predicts the force–intensity and force–frequency relationships after spinal cord injuries.
+    Muscle & Nerve: Official Journal of the American Association of Electrodiagnostic Medicine, 36(2), 214-222.
     """
 
     def __init__(
         self,
-        model_name: str = "hmed2018",
+        model_name: str = "ding_2007_with_fatigue",
         muscle_name: str = None,
         stim_time: list[float] = None,
         previous_stim: dict = None,
         sum_stim_truncation: int = 20,
     ):
-        if previous_stim:
-            if len(previous_stim["time"]) != len(previous_stim["pulse_intensity"]):
-                raise ValueError("The previous_stim time and pulse_intensity must have the same length")
-        super(DingModelPulseIntensityFrequency, self).__init__(
+        super(DingModelPulseWidthFrequencyWithFatigue, self).__init__(
             model_name=model_name,
             muscle_name=muscle_name,
             stim_time=stim_time,
             previous_stim=previous_stim,
             sum_stim_truncation=sum_stim_truncation,
         )
-        self._with_fatigue = False
-        self.stim_pulse_intensity_prev = []
-        self.previous_stim = previous_stim if previous_stim else {"time": [], "pulse_intensity": []}
+        self._with_fatigue = True
+        self.stim_time = stim_time
+        self.fmax = 315
 
-        # --- Default values ---#
-        AR_DEFAULT = 0.586  # (-) Translation of axis coordinates.
-        BS_DEFAULT = 0.026  # (-) Fiber muscle recruitment constant identification.
-        IS_DEFAULT = 63.1  # (mA) Muscle saturation intensity.
-        CR_DEFAULT = 0.833  # (-) Translation of axis coordinates.
+        # --- Default values --- #
+        ALPHA_A_DEFAULT = -4.0 * 10e-2  # Value from Ding's experimentation [1] (s^-2)
+        TAU_FAT_DEFAULT = 127  # Value from Ding's experimentation [1] (s)
+        ALPHA_TAU1_DEFAULT = 2.1 * 10e-6  # Value from Ding's experimentation [1] (N^-1)
+        ALPHA_KM_DEFAULT = 1.9 * 10e-6  # Value from Ding's experimentation [1] (s^-1.N^-1)
 
-        # ---- Custom values for the example ---- #
-        # ---- Force models ---- #
-        self.ar = AR_DEFAULT
-        self.bs = BS_DEFAULT
-        self.Is = IS_DEFAULT
-        self.cr = CR_DEFAULT
-        self.impulse_intensity = None
+        # ---- Fatigue models ---- #
+        self.alpha_a = ALPHA_A_DEFAULT
+        self.alpha_tau1 = ALPHA_TAU1_DEFAULT
+        self.tau_fat = TAU_FAT_DEFAULT
+        self.alpha_km = ALPHA_KM_DEFAULT
+
+    # ---- Absolutely needed methods ---- #
+    @property
+    def name_dof(self, with_muscle_name: bool = False) -> list[str]:
+        muscle_name = "_" + self.muscle_name if self.muscle_name and with_muscle_name else ""
+        return [
+            "Cn" + muscle_name,
+            "F" + muscle_name,
+            "A" + muscle_name,
+            "Tau1" + muscle_name,
+            "Km" + muscle_name,
+        ]
+
+    @property
+    def nb_state(self) -> int:
+        return 5
 
     @property
     def identifiable_parameters(self):
         return {
-            "a_rest": self.a_rest,
+            "a_scale": self.a_scale,
             "tau1_rest": self.tau1_rest,
             "km_rest": self.km_rest,
             "tau2": self.tau2,
-            "ar": self.ar,
-            "bs": self.bs,
-            "Is": self.Is,
-            "cr": self.cr,
+            "pd0": self.pd0,
+            "pdt": self.pdt,
+            "alpha_a": self.alpha_a,
+            "alpha_tau1": self.alpha_tau1,
+            "alpha_km": self.alpha_km,
+            "tau_fat": self.tau_fat,
         }
 
-    @property
-    def pulse_intensity_name(self):
-        muscle_name = "_" + self.muscle_name if self.muscle_name else ""
-        return "pulse_intensity" + muscle_name
+    def standard_rest_values(self) -> np.array:
+        """
+        Returns
+        -------
+        The rested values of Cn, F, A, Tau1, Km
+        """
+        return np.array([[0], [0], [self.a_scale], [self.tau1_rest], [self.km_rest]])
 
-    def set_ar(self, model, ar: MX | float):
-        # models is required for bioptim compatibility
-        self.ar = ar
-
-    def set_bs(self, model, bs: MX | float):
-        self.bs = bs
-
-    def set_Is(self, model, Is: MX | float):
-        self.Is = Is
-
-    def set_cr(self, model, cr: MX | float):
-        self.cr = cr
-
-    def get_lambda_i(self, nb_stim: int, pulse_intensity: MX | float) -> list[MX | float]:
-        return [self.lambda_i_calculation(pulse_intensity[i]) for i in range(nb_stim)]
-
-    # ---- Absolutely needed methods ---- #
     def serialize(self) -> tuple[Callable, dict]:
         # This is where you can serialize your models
         # This is useful if you want to save your models and load it later
         return (
-            DingModelPulseIntensityFrequency,
+            DingModelPulseWidthFrequencyWithFatigue,
             {
                 "tauc": self.tauc,
                 "a_rest": self.a_rest,
                 "tau1_rest": self.tau1_rest,
                 "km_rest": self.km_rest,
                 "tau2": self.tau2,
-                "ar": self.ar,
-                "bs": self.bs,
-                "Is": self.Is,
-                "cr": self.cr,
+                "alpha_a": self.alpha_a,
+                "alpha_tau1": self.alpha_tau1,
+                "alpha_km": self.alpha_km,
+                "tau_fat": self.tau_fat,
+                "a_scale": self.a_scale,
+                "pd0": self.pd0,
+                "pdt": self.pdt,
+                "stim_time": self.stim_time,
+                "previous_stim": self.previous_stim,
             },
         )
 
@@ -119,9 +123,12 @@ class DingModelPulseIntensityFrequency(DingModelFrequency):
         self,
         cn: MX,
         f: MX,
+        a: MX = None,
+        tau1: MX = None,
+        km: MX = None,
         t: MX = None,
-        t_stim_prev: list[MX] | list[float] = None,
-        pulse_intensity: list[MX] | list[float] = None,
+        t_stim_prev: MX = None,
+        pulse_width: MX = None,
         force_length_relationship: MX | float = 1,
         force_velocity_relationship: MX | float = 1,
         passive_force_relationship: MX | float = 0,
@@ -135,12 +142,18 @@ class DingModelPulseIntensityFrequency(DingModelFrequency):
             The value of the ca_troponin_complex (unitless)
         f: MX
             The value of the force (N)
+        a: MX
+            The value of the scaling factor (unitless)
+        tau1: MX
+            The value of the time_state_force_no_cross_bridge (ms)
+        km: MX
+            The value of the cross_bridges (unitless)
         t: MX
             The current time at which the dynamics is evaluated (s)
-        t_stim_prev: list[MX] | list[float]
-            The previous stimulation time (s)
-        pulse_intensity: list[MX] | list[float]
-            The pulsation intensity of the current stimulation (mA)
+        t_stim_prev: MX
+            The time of the previous stimulation (s)
+        pulse_width: MX
+            The time of the impulse (s)
         force_length_relationship: MX | float
             The force length relationship value (unitless)
         force_velocity_relationship: MX | float
@@ -152,68 +165,69 @@ class DingModelPulseIntensityFrequency(DingModelFrequency):
         -------
         The value of the derivative of each state dx/dt at the current time t
         """
-        cn_dot = self.calculate_cn_dot(cn, t, t_stim_prev, pulse_intensity)
+        cn_dot = self.calculate_cn_dot(cn, t, t_stim_prev)
+        a_scale = self.a_calculation(a_scale=a, pulse_width=pulse_width)
+
         f_dot = self.f_dot_fun(
             cn,
             f,
-            self.a_rest,
-            self.tau1_rest,
-            self.km_rest,
+            a_scale,
+            tau1,
+            km,
             force_length_relationship=force_length_relationship,
             force_velocity_relationship=force_velocity_relationship,
             passive_force_relationship=passive_force_relationship,
-        )  # Equation n°2
-        return vertcat(cn_dot, f_dot)
+        )  # Equation n°2 from Ding's 2003 article
 
-    def lambda_i_calculation(self, pulse_intensity: MX):
+        a_dot = self.a_dot_fun(a, f)
+        tau1_dot = self.tau1_dot_fun(tau1, f)  # Equation n°9 from Ding's 2003 article
+        km_dot = self.km_dot_fun(km, f)  # Equation n°11 from Ding's 2003 article
+        return vertcat(cn_dot, f_dot, a_dot, tau1_dot, km_dot)
+
+    def a_dot_fun(self, a: MX, f: MX) -> MX | float:
         """
         Parameters
         ----------
-        pulse_intensity: MX
-            The pulsation intensity of the current stimulation (mA)
+        a: MX
+            The previous step value of scaling factor (unitless)
+        f: MX
+            The previous step value of force (N)
+
         Returns
         -------
-        The lambda factor, part of the n°1 equation
+        The value of the derivative scaling factor (unitless)
         """
-        lambda_i = self.ar * (tanh(self.bs * (pulse_intensity - self.Is)) + self.cr)  # equation include intensity
-        return lambda_i
+        return -(a - self.a_scale) / self.tau_fat + self.alpha_a * f  # Equation n°5
 
-    @staticmethod
-    def lambda_i_calculation_identification(
-        pulse_intensity: MX, ar: MX | float, bs: MX | float, Is: MX | float, cr: MX | float
-    ):
+    def tau1_dot_fun(self, tau1: MX, f: MX) -> MX | float:
         """
         Parameters
         ----------
-        pulse_intensity: MX
-            The pulsation intensity of the current stimulation (mA)
-        ar: MX | float
-            Translation of axis coordinates (-)
-        bs: MX | float
-            Fiber muscle recruitment constant identification.
-        Is: MX | float
-            Muscle saturation intensity (mA)
-        cr: MX | float
-            Translation of axis coordinates (-)
+        tau1: MX
+            The previous step value of time_state_force_no_cross_bridge (ms)
+        f: MX
+            The previous step value of force (N)
+
         Returns
         -------
-        The lambda factor, part of the n°1 equation
+        The value of the derivative time_state_force_no_cross_bridge (ms)
         """
-        lambda_i = ar * (tanh(bs * (pulse_intensity - Is)) + cr)  # equation include intensity
-        return lambda_i
+        return -(tau1 - self.tau1_rest) / self.tau_fat + self.alpha_tau1 * f  # Equation n°9
 
-    def set_impulse_intensity(self, value: MX):
+    def km_dot_fun(self, km: MX, f: MX) -> MX | float:
         """
-        Sets the impulse intensity for each pulse (phases) according to the ocp parameter "impulse_intensity"
-
         Parameters
         ----------
-        value: MX
-            The pulsation intensity list (s)
+        km: MX
+            The previous step value of cross_bridges (unitless)
+        f: MX
+            The previous step value of force (N)
+
+        Returns
+        -------
+        The value of the derivative cross_bridges (unitless)
         """
-        self.impulse_intensity = []
-        for i in range(value.shape[0]):
-            self.impulse_intensity.append(value[i])
+        return -(km - self.km_rest) / self.tau_fat + self.alpha_km * f  # Equation n°11
 
     @staticmethod
     def dynamics(
@@ -224,7 +238,7 @@ class DingModelPulseIntensityFrequency(DingModelFrequency):
         algebraic_states: MX,
         numerical_timeseries: MX,
         nlp: NonLinearProgram,
-        fes_model: NonLinearProgram = None,
+        fes_model=None,
         force_length_relationship: MX | float = 1,
         force_velocity_relationship: MX | float = 1,
         passive_force_relationship: MX | float = 0,
@@ -248,7 +262,7 @@ class DingModelPulseIntensityFrequency(DingModelFrequency):
             The numerical timeseries of the system
         nlp: NonLinearProgram
             A reference to the phase
-        fes_model: DingModelPulseIntensityFrequency
+        fes_model: DingModelPulseWidthFrequencyWithFatigue
             The current phase fes model
         force_length_relationship: MX | float
             The force length relationship value (unitless)
@@ -267,9 +281,12 @@ class DingModelPulseIntensityFrequency(DingModelFrequency):
             dxdt=dxdt_fun(
                 cn=states[0],
                 f=states[1],
+                a=states[2],
+                tau1=states[3],
+                km=states[4],
                 t=time,
                 t_stim_prev=numerical_timeseries,
-                pulse_intensity=controls,
+                pulse_width=controls[0],
                 force_length_relationship=force_length_relationship,
                 force_velocity_relationship=force_velocity_relationship,
                 passive_force_relationship=passive_force_relationship,
@@ -299,20 +316,5 @@ class DingModelPulseIntensityFrequency(DingModelFrequency):
             The type of contact to use for the model
         """
         StateConfigure().configure_all_fes_model_states(ocp, nlp, fes_model=self)
-        StateConfigure().configure_pulse_intensity(ocp, nlp, truncation=self.sum_stim_truncation)
+        StateConfigure().configure_last_pulse_width(ocp, nlp)
         ConfigureProblem.configure_dynamics_function(ocp, nlp, dyn_func=self.dynamics)
-
-    def min_pulse_intensity(self):
-        """
-        Returns
-        -------
-        The minimum pulse intensity threshold of the model
-        For lambda_i = ar * (tanh(bs * (pulse_intensity - Is)) + cr) > 0
-        """
-        return (np.arctanh(-self.cr) / self.bs) + self.Is
-
-    def _get_additional_previous_stim_time(self):
-        while len(self.previous_stim["time"]) < self.sum_stim_truncation:
-            self.previous_stim["time"].insert(0, -10000000)
-            self.previous_stim["pulse_intensity"].insert(0, 50)
-        return self.previous_stim
