+++ cocofest/models/hmed2018/hmed2018.py
@@ -1,41 +1,43 @@
 from typing import Callable
 
+from casadi import MX, vertcat, tanh
 import numpy as np
-from casadi import MX, vertcat, exp
 
 from bioptim import (
     ConfigureProblem,
     DynamicsEvaluation,
     NonLinearProgram,
     OptimalControlProgram,
-    ParameterList,
 )
-from .ding2003 import DingModelFrequency
-from .state_configure import StateConfigure
+from cocofest.models.ding2003.ding2003 import DingModelFrequency
+from cocofest.models.state_configure import StateConfigure
 
 
-class DingModelPulseWidthFrequency(DingModelFrequency):
+class DingModelPulseIntensityFrequency(DingModelFrequency):
     """
     This is a custom models that inherits from bioptim. CustomModel.
     As CustomModel is an abstract class, some methods are mandatory and must be implemented.
     Such as serialize, name_dof, nb_state.
 
-    This is the Ding 2007 model using the stimulation frequency and pulse width in input.
+    This is the Hmed 2018 model using the stimulation frequency and pulse intensity in input.
 
-    Ding, J., Chou, L. W., Kesar, T. M., Lee, S. C., Johnston, T. E., Wexler, A. S., & Binder‐Macleod, S. A. (2007).
-    Mathematical model that predicts the force–intensity and force–frequency relationships after spinal cord injuries.
-    Muscle & Nerve: Official Journal of the American Association of Electrodiagnostic Medicine, 36(2), 214-222.
+    Hmed, A. B., Bakir, T., Garnier, Y. M., Sakly, A., Lepers, R., & Binczak, S. (2018).
+    An approach to a muscle force model with force-pulse amplitude relationship of human quadriceps muscles.
+    Computers in Biology and Medicine, 101, 218-228.
     """
 
     def __init__(
         self,
-        model_name: str = "ding_2007",
+        model_name: str = "hmed2018",
         muscle_name: str = None,
         stim_time: list[float] = None,
         previous_stim: dict = None,
         sum_stim_truncation: int = 20,
     ):
-        super(DingModelPulseWidthFrequency, self).__init__(
+        if previous_stim:
+            if len(previous_stim["time"]) != len(previous_stim["pulse_intensity"]):
+                raise ValueError("The previous_stim time and pulse_intensity must have the same length")
+        super(DingModelPulseIntensityFrequency, self).__init__(
             model_name=model_name,
             muscle_name=muscle_name,
             stim_time=stim_time,
@@ -43,67 +45,73 @@ class DingModelPulseWidthFrequency(DingModelFrequency):
             sum_stim_truncation=sum_stim_truncation,
         )
         self._with_fatigue = False
-        self.pulse_width = None
-        self.previous_stim = previous_stim if previous_stim else {"time": []}
-        self.stim_time = stim_time
+        self.stim_pulse_intensity_prev = []
+        self.previous_stim = previous_stim if previous_stim else {"time": [], "pulse_intensity": []}
 
-        # --- Default values --- #
-        A_SCALE_DEFAULT = 4920  # Value from Ding's 2007 article (N/s)
-        PD0_DEFAULT = 0.000131405  # Value from Ding's 2007 article (s)
-        PDT_DEFAULT = 0.000194138  # Value from Ding's 2007 article (s)
-        TAU1_REST_DEFAULT = 0.060601  # Value from Ding's 2003 article (s)
-        TAU2_DEFAULT = 0.001  # Value from Ding's 2007 article (s)
-        KM_REST_DEFAULT = 0.137  # Value from Ding's 2007 article (unitless)
-        TAUC_DEFAULT = 0.011  # Value from Ding's 2007 article (s)
+        # --- Default values ---#
+        AR_DEFAULT = 0.586  # (-) Translation of axis coordinates.
+        BS_DEFAULT = 0.026  # (-) Fiber muscle recruitment constant identification.
+        IS_DEFAULT = 63.1  # (mA) Muscle saturation intensity.
+        CR_DEFAULT = 0.833  # (-) Translation of axis coordinates.
 
         # ---- Custom values for the example ---- #
         # ---- Force models ---- #
-        self.a_scale = A_SCALE_DEFAULT
-        self.pd0 = PD0_DEFAULT
-        self.pdt = PDT_DEFAULT
-        self.tau1_rest = TAU1_REST_DEFAULT
-        self.tau2 = TAU2_DEFAULT
-        self.km_rest = KM_REST_DEFAULT
-        self.tauc = TAUC_DEFAULT
-        self.fmax = 248  # Maximum force (N) at 100 Hz and 600 us
+        self.ar = AR_DEFAULT
+        self.bs = BS_DEFAULT
+        self.Is = IS_DEFAULT
+        self.cr = CR_DEFAULT
+        self.impulse_intensity = None
 
     @property
     def identifiable_parameters(self):
         return {
-            "a_scale": self.a_scale,
+            "a_rest": self.a_rest,
             "tau1_rest": self.tau1_rest,
             "km_rest": self.km_rest,
             "tau2": self.tau2,
-            "pd0": self.pd0,
-            "pdt": self.pdt,
+            "ar": self.ar,
+            "bs": self.bs,
+            "Is": self.Is,
+            "cr": self.cr,
         }
 
-    def set_a_scale(self, model, a_scale: MX | float):
+    @property
+    def pulse_intensity_name(self):
+        muscle_name = "_" + self.muscle_name if self.muscle_name else ""
+        return "pulse_intensity" + muscle_name
+
+    def set_ar(self, model, ar: MX | float):
         # models is required for bioptim compatibility
-        self.a_scale = a_scale
+        self.ar = ar
 
-    def set_pd0(self, model, pd0: MX | float):
-        self.pd0 = pd0
+    def set_bs(self, model, bs: MX | float):
+        self.bs = bs
 
-    def set_pdt(self, model, pdt: MX | float):
-        self.pdt = pdt
+    def set_Is(self, model, Is: MX | float):
+        self.Is = Is
 
+    def set_cr(self, model, cr: MX | float):
+        self.cr = cr
+
+    def get_lambda_i(self, nb_stim: int, pulse_intensity: MX | float) -> list[MX | float]:
+        return [self.lambda_i_calculation(pulse_intensity[i]) for i in range(nb_stim)]
+
+    # ---- Absolutely needed methods ---- #
     def serialize(self) -> tuple[Callable, dict]:
         # This is where you can serialize your models
         # This is useful if you want to save your models and load it later
         return (
-            DingModelPulseWidthFrequency,
+            DingModelPulseIntensityFrequency,
             {
                 "tauc": self.tauc,
                 "a_rest": self.a_rest,
                 "tau1_rest": self.tau1_rest,
                 "km_rest": self.km_rest,
                 "tau2": self.tau2,
-                "a_scale": self.a_scale,
-                "pd0": self.pd0,
-                "pdt": self.pdt,
-                "stim_time": self.stim_time,
-                "previous_stim": self.previous_stim,
+                "ar": self.ar,
+                "bs": self.bs,
+                "Is": self.Is,
+                "cr": self.cr,
             },
         )
 
@@ -112,8 +120,8 @@ class DingModelPulseWidthFrequency(DingModelFrequency):
         cn: MX,
         f: MX,
         t: MX = None,
-        t_stim_prev: list[float] | list[MX] = None,
-        pulse_width: MX = None,
+        t_stim_prev: list[MX] | list[float] = None,
+        pulse_intensity: list[MX] | list[float] = None,
         force_length_relationship: MX | float = 1,
         force_velocity_relationship: MX | float = 1,
         passive_force_relationship: MX | float = 0,
@@ -129,88 +137,83 @@ class DingModelPulseWidthFrequency(DingModelFrequency):
             The value of the force (N)
         t: MX
             The current time at which the dynamics is evaluated (s)
-        t_stim_prev: list[float] | list[MX]
-            The time list of the previous stimulations (s)
-        pulse_width: MX
-            The pulsation duration of the current stimulation (s)
+        t_stim_prev: list[MX] | list[float]
+            The previous stimulation time (s)
+        pulse_intensity: list[MX] | list[float]
+            The pulsation intensity of the current stimulation (mA)
         force_length_relationship: MX | float
             The force length relationship value (unitless)
         force_velocity_relationship: MX | float
             The force velocity relationship value (unitless)
         passive_force_relationship: MX | float
-            The passive force coefficient of the muscle (unitless)
+            The passive force relationship value (unitless)
 
         Returns
         -------
         The value of the derivative of each state dx/dt at the current time t
         """
-        cn_dot = self.calculate_cn_dot(cn, t, t_stim_prev)
-        a_scale = self.a_calculation(a_scale=self.a_scale, pulse_width=pulse_width)
+        cn_dot = self.calculate_cn_dot(cn, t, t_stim_prev, pulse_intensity)
         f_dot = self.f_dot_fun(
             cn,
             f,
-            a_scale,
+            self.a_rest,
             self.tau1_rest,
             self.km_rest,
             force_length_relationship=force_length_relationship,
             force_velocity_relationship=force_velocity_relationship,
             passive_force_relationship=passive_force_relationship,
-        )  # Equation n°2 from Ding's 2003 article
+        )  # Equation n°2
         return vertcat(cn_dot, f_dot)
 
-    def a_calculation(
-        self,
-        a_scale: float | MX,
-        pulse_width: MX,
-    ) -> MX:
+    def lambda_i_calculation(self, pulse_intensity: MX):
         """
         Parameters
         ----------
-        a_scale: float | MX
-            The scaling factor of the current stimulation (unitless)
-        pulse_width: MX
-            The pulsation duration of the current stimulation (s)
+        pulse_intensity: MX
+            The pulsation intensity of the current stimulation (mA)
         Returns
         -------
-        The value of scaling factor (unitless)
+        The lambda factor, part of the n°1 equation
         """
-        return a_scale * (1 - exp(-(pulse_width - self.pd0) / self.pdt))
+        lambda_i = self.ar * (tanh(self.bs * (pulse_intensity - self.Is)) + self.cr)  # equation include intensity
+        return lambda_i
 
-    def a_calculation_identification(
-        self,
-        a_scale: float | MX,
-        pulse_width: MX,
-        pd0: float | MX,
-        pdt: float | MX,
-    ) -> MX:
+    @staticmethod
+    def lambda_i_calculation_identification(
+        pulse_intensity: MX, ar: MX | float, bs: MX | float, Is: MX | float, cr: MX | float
+    ):
         """
         Parameters
         ----------
-        a_scale: float | MX
-            The scaling factor of the current stimulation (unitless)
-        pulse_width: MX
-            The pulsation duration of the current stimulation (s)
-        pd0: float | MX
-            The pd0 value (s)
-        pdt: float | MX
-            The pdt value (s)
-
+        pulse_intensity: MX
+            The pulsation intensity of the current stimulation (mA)
+        ar: MX | float
+            Translation of axis coordinates (-)
+        bs: MX | float
+            Fiber muscle recruitment constant identification.
+        Is: MX | float
+            Muscle saturation intensity (mA)
+        cr: MX | float
+            Translation of axis coordinates (-)
         Returns
         -------
-        The value of scaling factor (unitless)
+        The lambda factor, part of the n°1 equation
         """
-        return a_scale * (1 - exp(-(pulse_width - pd0) / pdt))
+        lambda_i = ar * (tanh(bs * (pulse_intensity - Is)) + cr)  # equation include intensity
+        return lambda_i
 
-    def set_impulse_width(self, value: list[MX]):
+    def set_impulse_intensity(self, value: MX):
         """
-        Sets the pulse width for each pulse (phases) according to the ocp parameter "pulse_width"
+        Sets the impulse intensity for each pulse (phases) according to the ocp parameter "impulse_intensity"
 
         Parameters
         ----------
-        value: list[MX]
-            The pulsation duration list (s)
+        value: MX
+            The pulsation intensity list (s)
         """
-        self.pulse_width = value
+        self.impulse_intensity = []
+        for i in range(value.shape[0]):
+            self.impulse_intensity.append(value[i])
 
     @staticmethod
     def dynamics(
@@ -221,7 +224,7 @@ class DingModelPulseWidthFrequency(DingModelFrequency):
         algebraic_states: MX,
         numerical_timeseries: MX,
         nlp: NonLinearProgram,
-        fes_model=None,
+        fes_model: NonLinearProgram = None,
         force_length_relationship: MX | float = 1,
         force_velocity_relationship: MX | float = 1,
         passive_force_relationship: MX | float = 0,
@@ -245,14 +248,14 @@ class DingModelPulseWidthFrequency(DingModelFrequency):
             The numerical timeseries of the system
         nlp: NonLinearProgram
             A reference to the phase
-        fes_model: DingModelPulseWidthFrequency
+        fes_model: DingModelPulseIntensityFrequency
             The current phase fes model
         force_length_relationship: MX | float
             The force length relationship value (unitless)
         force_velocity_relationship: MX | float
             The force velocity relationship value (unitless)
         passive_force_relationship: MX | float
-            The passive force coefficient of the muscle (unitless)
+            The passive force relationship value (unitless)
         Returns
         -------
         The derivative of the states in the tuple[MX] format
@@ -266,7 +269,7 @@ class DingModelPulseWidthFrequency(DingModelFrequency):
                 f=states[1],
                 t=time,
                 t_stim_prev=numerical_timeseries,
-                pulse_width=controls[0],
+                pulse_intensity=controls,
                 force_length_relationship=force_length_relationship,
                 force_velocity_relationship=force_velocity_relationship,
                 passive_force_relationship=passive_force_relationship,
@@ -296,5 +299,20 @@ class DingModelPulseWidthFrequency(DingModelFrequency):
             The type of contact to use for the model
         """
         StateConfigure().configure_all_fes_model_states(ocp, nlp, fes_model=self)
-        StateConfigure().configure_last_pulse_width(ocp, nlp)
+        StateConfigure().configure_pulse_intensity(ocp, nlp, truncation=self.sum_stim_truncation)
         ConfigureProblem.configure_dynamics_function(ocp, nlp, dyn_func=self.dynamics)
+
+    def min_pulse_intensity(self):
+        """
+        Returns
+        -------
+        The minimum pulse intensity threshold of the model
+        For lambda_i = ar * (tanh(bs * (pulse_intensity - Is)) + cr) > 0
+        """
+        return (np.arctanh(-self.cr) / self.bs) + self.Is
+
+    def _get_additional_previous_stim_time(self):
+        while len(self.previous_stim["time"]) < self.sum_stim_truncation:
+            self.previous_stim["time"].insert(0, -10000000)
+            self.previous_stim["pulse_intensity"].insert(0, 50)
+        return self.previous_stim
