+++ cocofest/models/marion2009/marion2009_modified.py
@@ -1,16 +1,12 @@
 from typing import Callable
 
-import numpy as np
 from casadi import MX, vertcat
 
 from bioptim import (
-    ConfigureProblem,
     DynamicsEvaluation,
     NonLinearProgram,
-    OptimalControlProgram,
 )
-from .ding2007 import DingModelPulseWidthFrequency
-from .state_configure import StateConfigure
+from cocofest.models.ding2007.ding2007 import DingModelPulseWidthFrequency
 
 
 class Marion2009ModelPulseWidthFrequency(DingModelPulseWidthFrequency):
