+++ cocofest/models/marion2009/marion2009_with_fatigue.py
@@ -7,7 +7,7 @@ from bioptim import (
     DynamicsEvaluation,
     NonLinearProgram,
 )
-from .marion2009 import Marion2009ModelFrequency
+from cocofest.models.marion2009 import Marion2009ModelFrequency
 
 
 class Marion2009ModelFrequencyWithFatigue(Marion2009ModelFrequency):
