+++ cocofest/models/model_maker.py
@@ -1,17 +1,17 @@
-from .ding2003 import DingModelFrequency
-from .ding2003_with_fatigue import DingModelFrequencyWithFatigue
-from .ding2007 import DingModelPulseWidthFrequency
-from .ding2007_with_fatigue import DingModelPulseWidthFrequencyWithFatigue
-from .hmed2018 import DingModelPulseIntensityFrequency
-from .hmed2018_with_fatigue import DingModelPulseIntensityFrequencyWithFatigue
-from .marion2009 import Marion2009ModelFrequency
-from .marion2009_with_fatigue import Marion2009ModelFrequencyWithFatigue
-from .marion2009_modified import Marion2009ModelPulseWidthFrequency
-from .marion2009_modified_with_fatigue import Marion2009ModelPulseWidthFrequencyWithFatigue
-from .marion2013 import Marion2013ModelFrequency
-from .marion2013_with_fatigue import Marion2013ModelFrequencyWithFatigue
-from .marion2013_modified import Marion2013ModelPulseWidthFrequency
-from .marion2013_modified_with_fatigue import Marion2013ModelPulseWidthFrequencyWithFatigue
+from cocofest.models.ding2003.ding2003 import DingModelFrequency
+from cocofest.models.ding2003.ding2003_with_fatigue import DingModelFrequencyWithFatigue
+from cocofest.models.ding2007.ding2007 import DingModelPulseWidthFrequency
+from cocofest.models.ding2007.ding2007_with_fatigue import DingModelPulseWidthFrequencyWithFatigue
+from cocofest.models.hmed2018.hmed2018 import DingModelPulseIntensityFrequency
+from cocofest.models.hmed2018.hmed2018_with_fatigue import DingModelPulseIntensityFrequencyWithFatigue
+from cocofest.models.marion2009.marion2009 import Marion2009ModelFrequency
+from cocofest.models.marion2009.marion2009_with_fatigue import Marion2009ModelFrequencyWithFatigue
+from cocofest.models.marion2009.marion2009_modified import Marion2009ModelPulseWidthFrequency
+from cocofest.models.marion2009.marion2009_modified_with_fatigue import Marion2009ModelPulseWidthFrequencyWithFatigue
+from cocofest.models.marion2013.marion2013 import Marion2013ModelFrequency
+from cocofest.models.marion2013.marion2013_with_fatigue import Marion2013ModelFrequencyWithFatigue
+from cocofest.models.marion2013.marion2013_modified import Marion2013ModelPulseWidthFrequency
+from cocofest.models.marion2013.marion2013_modified_with_fatigue import Marion2013ModelPulseWidthFrequencyWithFatigue
 
 
 class ModelMaker:
