+++ examples/fes_model_comparision/marion2013_example.py
@@ -7,7 +7,7 @@ end of the last node while minimizing the muscle force state.
 import numpy as np
 from bioptim import Solver, ObjectiveList, ObjectiveFcn, OptimalControlProgram, ControlType, Node, BoundsList, InitialGuessList
 from cocofest import OcpFes, ModelMaker, FES_plot
-from cocofest.models.marion2013_modified import Marion2013ModelPulseWidthFrequency
+from cocofest.models.marion2013.marion2013_modified import Marion2013ModelPulseWidthFrequency
 
 
 def prepare_ocp(model, final_time, pw_max=0.0006):
